requirements.txt
fastapi
uvicorn
pydantic
supabase
python-multipart
python-dotenv
moviepy
openai-whisper
pydub
librosa
numpy
openai
google-generativeai
opencv-python
mediapipe
